requirements.txt
streamlit
transformers
torch
pillow
pinecone-client
python-dotenv
google-generativeai
requests